Add v1alpha1_to_v2.go
